Create Code_file_before_new_code_from_mac.html
Create Code_file_before_new_code_from_mac.html
In this i have modified the code where the initially uploaded file was not playing, but i think that was due to the wrong format of the uploaded csv file. but this code is working, if ever wanted to erevrt than use this code as the base code.