Update index.html
Blocked the Right click option to inspect the source code as a security concern of code copyright. Below is exact code added newly at last part of code:
JavaScript

Just have added the below block of code in the best code version of v7, remaining all functionality is as it was in v7. Keep it carefully & safe for future refrences.


/* Blocked the Right click option to inspect the source code as a security concern of code copyright.
Below is exact code added newly at last part of code: */
        document.addEventListener('contextmenu', event => event.preventDefault());
        document.addEventListener('keydown', e => {
            if (e.ctrlKey && (e.key === 'u' || e.key === 'U' || e.key === 'i' || e.key === 'I' || e.key === 'j' || e.key === 'J')) {
                e.preventDefault();
            }
            if (e.keyCode === 123) { // F12 key
                e.preventDefault();
            }
        });